+++ internal/storage/postgres/storage.go
@@ -7,9 +7,9 @@ import (
 	"time"
 
 	"github.com/jmoiron/sqlx"
+	server "github.com/mrkovshik/yandex_diploma/api"
 	"github.com/mrkovshik/yandex_diploma/internal/apperrors"
 	"github.com/mrkovshik/yandex_diploma/internal/model"
-	server "github.com/mrkovshik/yandex_diploma/internal/service/loyalty"
 )
 
 type Storage struct {
